Move zoning district model computeds into component
