+++ app/templates/components/layer-group-display/zoning-district.hbs
@@ -1,11 +1,36 @@
-<label class="header-type">ZONING DISTRICT</label>
-<h1 class="header-large">{{model.value.id}}</h1>
-<p>{{model.value.description}}</p>
+<label class="header-type">
+  ZONING DISTRICT
+</label>
+
+<h1 class="header-large">
+  {{model.value.id}}
+</h1>
+
+<p>
+  {{description}}
+</p>
+
 {{#unless (eq model.value.id 'BPC')}}
-  <p><a href="https://www1.nyc.gov/site/planning/zoning/districts-tools/{{primaryzoneURL}}.page" target="_blank">{{fa-icon 'external-link-alt'}} Learn more about {{model.value.id}} districts&hellip;</a></p>
-  <button {{action "fitBounds"}} class="button tiny hollow">Fit map to all {{model.value.id}} districts</button>
+  <p>
+    <a 
+      href="https://www1.nyc.gov/site/planning/zoning/districts-tools/{{primaryzoneURL}}.page" 
+      target="_blank"
+    >
+      {{fa-icon 'external-link-alt'}} Learn more about {{id}} districts&hellip;</a></p>
+  <button {{action "fitBounds"}}
+    class="button tiny hollow"
+  >
+    Fit map to all {{model.value.id}} districts
+  </button>
 {{else}}
-  <p><a href="https://www1.nyc.gov/site/planning/zoning/districts-tools/special-purpose-districts-manhattan.page#battery_park" target="_blank">{{fa-icon 'external-link-alt'}} Learn more about the Special Battery Park City District&hellip;</a></p>
+  <p>
+    <a
+      href="https://www1.nyc.gov/site/planning/zoning/districts-tools/special-purpose-districts-manhattan.page#battery_park"
+      target="_blank"
+    >
+      {{fa-icon 'external-link-alt'}} Learn more about the Special Battery Park City District&hellip;
+    </a>
+  </p>
 {{/unless}}
 
 {{outlet}}
